+++ tests/TestCase/Core/PluginTest.php
@@ -67,6 +67,7 @@ class PluginTest extends TestCase
         $this->assertSame(ROOT . $file, Plugin::path('MeTools', $file));
         $this->assertSame(ROOT . $file, Plugin::path('MeTools', $file, true));
 
-        $this->assertException(fn() => Plugin::path('MeTools', 'no_existing.php', true), FileNotExistsException::class, 'File or directory `MeTools.no_existing.php` does not exist');
+        $this->expectExceptionMessage('File or directory `MeTools.no_existing.php` does not exist');
+        Plugin::path('MeTools', 'no_existing.php', true);
     }
 }